api/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
albumentations>=1.3.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-json-logger==2.0.7
mangum>=0.17.0
awslambdaric>=2.0.0
boto3>=1.28.0